split opts into their own file
